refactor: Remove all implementation logic from algo-ts
