+++ packages/algo-ts/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@algorandfoundation/algorand-typescript",
-  "version": "1.0.0-beta.15",
+  "version": "1.0.0-beta.16",
   "description": "This package contains definitions for the types which comprise Algorand TypeScript which can be compiled to run on the Algorand Virtual Machine using the Puya compiler.",
   "private": false,
   "main": "index.js",
@@ -18,7 +18,7 @@
     "build:4-copy-pkg-json": "tstk copy-package-json -c",
     "build:5-copy-readme": "copyfiles ./readme.md ./dist",
     "watch": "rollup -c -w --configPlugin typescript",
-    "test": "vitest run"
+    "test": "echo 'No tests'"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
