--- packages/algo-ts/src/impl/name-of-type.ts
@@ -1,10 +1,0 @@
-export const nameOfType = (x: unknown) => {
-  if (typeof x === 'object') {
-    if (x === null) return 'Null'
-    if (x === undefined) return 'undefined'
-    if ('constructor' in x) {
-      return x.constructor.name
-    }
-  }
-  return typeof x
-}
