+++ packages/algo-ts/src/internal/typescript-helpers.ts
