+++ packages/algo-ts/src/op.ts
@@ -1,6 +1,6 @@
 /* THIS FILE IS GENERATED BY ~/scripts/generate-op-funcs.ts - DO NOT MODIFY DIRECTLY */
-import { NoImplementation } from './impl/errors'
-import { bytes, uint64, biguint } from './primitives'
+import { NoImplementation } from './internal/errors'
+import { biguint, bytes, uint64 } from './primitives'
 import { Account, Application, Asset } from './reference'
 
 export enum Base64 {
