+++ packages/algo-ts/tsconfig.build.json
@@ -6,6 +6,5 @@
     "declaration": true,
     "stripInternal": true
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.spec.ts", "src/**/*.test.ts", "src/**/tests/**"]
+  "include": ["src/**/*.ts"]
 }
