+++ packages/algo-ts/tsconfig.json
@@ -8,5 +8,5 @@
     "resolveJsonModule": true,
     "esModuleInterop": true
   },
-  "include": ["src/**/*.ts","scripts/**/*.ts", "examples/**/*.ts", "rollup.config.ts"]
+  "include": ["src/**/*.ts","rollup.config.ts"]
 }
